Add gRPC service for Catalog command
This commit introduces a new gRPC service for Catalog command operations. It includes the definition of several server-side message handlers, such as CreateCatalog, DeleteCatalog, and ActivateCatalog. Additionally, changes to the integration layer were made to enable the new gRPC service to interact with the current application's infrastructure, including the event store and event bus. A new migration for the event store database was also included.