Add EventBus infrastructure for Notification Service
This commit introduces the EventBus infrastructure for handling the Notification Service. Included are foundational elements such as the EventBus options, custom filters and observers for logging and validation, as well as the EventBus Gateway. Consumer classes for event handling are also defined. The new infrastructure will enhance system-wide message communication and improve the organization as well as error handling within the service.