Remove redundant HttpClients and ViewModels
This commit deletes a host of extraneous HttpClient and ViewModel files, replacing them with more focused API and Redux-based state management systems in the `Store/Cataloging` directory. It's part of an ongoing effort to simplify and flatten our codebase, removing unnecessary layers of complexity and improving maintainability. The replacement systems offer equivalent functionality, but with a more streamlined, easy-to-understand structure.